services/paddle-ocr/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
paddleocr==2.8.1
paddlepaddle==2.6.1
pillow==11.0.0
